+++ stable_baselines/acktr/acktr_disc.py
@@ -343,7 +343,7 @@ class ACKTR(ActorCriticRLModel):
 
         return self
 
-    def save(self, save_path):
+    def save(self, save_path, cloudpickle=False):
         data = {
             "gamma": self.gamma,
             "nprocs": self.nprocs,
@@ -366,4 +366,4 @@ class ACKTR(ActorCriticRLModel):
 
         params_to_save = self.get_parameters()
 
-        self._save_to_file(save_path, data=data, params=params_to_save)
+        self._save_to_file(save_path, data=data, params=params_to_save, cloudpickle=cloudpickle)