+++ stable_baselines/ppo2/ppo2.py
@@ -401,7 +401,7 @@ class PPO2(ActorCriticRLModel):
 
             return self
 
-    def save(self, save_path):
+    def save(self, save_path, cloudpickle=False):
         data = {
             "gamma": self.gamma,
             "n_steps": self.n_steps,
@@ -425,7 +425,7 @@ class PPO2(ActorCriticRLModel):
 
         params_to_save = self.get_parameters()
 
-        self._save_to_file(save_path, data=data, params=params_to_save)
+        self._save_to_file(save_path, data=data, params=params_to_save, cloudpickle=cloudpickle)
 
 
 class Runner(AbstractEnvRunner):