+++ stable_baselines/sac/sac.py
@@ -526,7 +526,7 @@ class SAC(OffPolicyRLModel):
         return (self.params +
                 self.target_params)
 
-    def save(self, save_path):
+    def save(self, save_path, cloudpickle=False):
         data = {
             "learning_rate": self.learning_rate,
             "buffer_size": self.buffer_size,
@@ -554,4 +554,4 @@ class SAC(OffPolicyRLModel):
 
         params_to_save = self.get_parameters()
 
-        self._save_to_file(save_path, data=data, params=params_to_save)
+        self._save_to_file(save_path, data=data, params=params_to_save, cloudpickle=cloudpickle)