+++ tests/test_continuous.py
@@ -59,7 +59,7 @@ def test_model_manipulation(request, model_class):
         acc_reward = sum(acc_reward) / N_TRIALS
 
         # saving
-        model_fname = './test_model_{}.pkl'.format(request.node.name)
+        model_fname = './test_model_{}.zip'.format(request.node.name)
         model.save(model_fname)
 
         del model, env
@@ -144,8 +144,8 @@ def test_model_manipulation(request, model_class):
         del model, env
 
     finally:
-        if os.path.exists("./test_model"):
-            os.remove("./test_model")
+        if os.path.exists("./test_model.zip"):
+            os.remove("./test_model.zip")
 
 
 def test_ddpg():
@@ -178,9 +178,9 @@ def test_ddpg_normalization():
     model.learn(1000)
     obs_rms_params = model.sess.run(model.obs_rms_params)
     ret_rms_params = model.sess.run(model.ret_rms_params)
-    model.save('./test_ddpg.pkl')
+    model.save('./test_ddpg.zip')
 
-    loaded_model = DDPG.load('./test_ddpg.pkl')
+    loaded_model = DDPG.load('./test_ddpg.zip')
     obs_rms_params_2 = loaded_model.sess.run(loaded_model.obs_rms_params)
     ret_rms_params_2 = loaded_model.sess.run(loaded_model.ret_rms_params)
 
@@ -190,8 +190,8 @@ def test_ddpg_normalization():
 
     del model, loaded_model
 
-    if os.path.exists("./test_ddpg.pkl"):
-        os.remove("./test_ddpg.pkl")
+    if os.path.exists("./test_ddpg.zip"):
+        os.remove("./test_ddpg.zip")
 
 
 def test_ddpg_popart():