+++ tests/test_save.py
@@ -1,5 +1,7 @@
 import os
 from io import BytesIO
+import json
+import zipfile
 
 import pytest
 import numpy as np
@@ -8,6 +10,7 @@ from stable_baselines import A2C, ACER, ACKTR, DQN, PPO1, PPO2, TRPO
 from stable_baselines.common import set_global_seeds
 from stable_baselines.common.identity_env import IdentityEnv
 from stable_baselines.common.vec_env import DummyVecEnv
+from stable_baselines.common.policies import MlpPolicy, FeedForwardPolicy
 
 N_TRIALS = 2000
 
@@ -26,19 +29,29 @@ STORE_METHODS = [
     "file-like"
 ]
 
+STORE_FORMAT = [
+    "zip",
+    "cloudpickle"
+]
 
 @pytest.mark.slow
 @pytest.mark.parametrize("model_class", MODEL_LIST)
 @pytest.mark.parametrize("storage_method", STORE_METHODS)
-def test_model_manipulation(request, model_class, storage_method):
+@pytest.mark.parametrize("store_format", STORE_FORMAT)
+def test_model_manipulation(request, model_class, storage_method, store_format):
     """
     Test if the algorithm (with a given policy) can be loaded and saved without any issues, the environment switching
     works and that the action prediction works
 
     :param model_class: (BaseRLModel) A RL model
+    :param storage_method: (str) Should file be saved to a file ("path") or to a buffer 
+        ("file-like")
+    :param store_format: (str) Save format, either "zip" or "cloudpickle".
     """
-
-    model_fname = './test_model_{}.pkl'.format(request.node.name)
+    
+    # Use postfix ".model" so we can remove the file later
+    model_fname = './test_model_{}.model'.format(request.node.name)
+    store_as_cloudpickle = store_format == "cloudpickle"
 
     try:
         env = DummyVecEnv([lambda: IdentityEnv(10)])
@@ -72,10 +85,10 @@ def test_model_manipulation(request, model_class, storage_method):
 
         # saving
         if storage_method == "path":  # saving to a path
-            model.save(model_fname)
+            model.save(model_fname, cloudpickle=store_as_cloudpickle)
         else:  # saving to a file-like object (BytesIO in this case)
             b_io = BytesIO()
-            model.save(b_io)
+            model.save(b_io, cloudpickle=store_as_cloudpickle)
             model_bytes = b_io.getvalue()
             b_io.close()
 
@@ -127,6 +140,79 @@ def test_model_manipulation(request, model_class, storage_method):
             obs, _, _, _ = env.step(action)
 
         del model, env
+
+    finally:
+        if os.path.exists(model_fname):
+            os.remove(model_fname)
+
+class CustomMlpPolicy(FeedForwardPolicy):
+    """A dummy "custom" policy to test out custom_objects"""
+    def __init__(self, sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse=False, **_kwargs):
+        super(CustomMlpPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps,
+                                              n_batch, reuse, feature_extraction="mlp",
+                                              **_kwargs)
+
+
+@pytest.mark.parametrize("model_class", MODEL_LIST)
+def test_save_custom_objects(request, model_class):
+    """
+    Test feeding custom_objects in model.load(...) function
+    """
+    # Skip DQN (not an actor-critic policy)
+    if model_class == DQN:
+        return
+
+    model_fname = './test_model_{}.zip'.format(request.node.name)
+
+    try:
+        env = DummyVecEnv([lambda: IdentityEnv(10)])
+
+        # Create and save model with default MLP policy
+        model = model_class(policy=MlpPolicy, env=env)
+        model.save(model_fname)
+
+        del model, env
+
+        # Corrupt "policy" serialization in the file
+        data_file = zipfile.ZipFile(model_fname, "r")
+        # Load all data (can't just update one file in the archive)
+        parameter_list = data_file.read("parameter_list")
+        parameters = data_file.read("parameters")
+        class_data = json.loads(data_file.read("data").decode())
+        data_file.close()
+
+        # Corrupt serialization of the "policy"
+        class_data["policy"][":serialized:"] = (
+            "Adding this should break serialization" +
+            class_data["policy"][":serialized:"]
+        )
+
+        # And dump everything back to the model file
+        data_file = zipfile.ZipFile(model_fname, "w")
+        data_file.writestr("data", json.dumps(class_data))
+        data_file.writestr("parameter_list", parameter_list)
+        data_file.writestr("parameters", parameters)
+        data_file.close()
+
+        # Try loading the model. This should
+        # result in an error
+        with pytest.raises(RuntimeError):
+            model = model_class.load(model_fname)
+
+        # Load model with custom objects ("custom" MlpPolicy)
+        # and it should work fine.
+        # Note: We could load model with just vanilla
+        #       MlpPolicy, too.
+        model = model_class.load(
+            model_fname,
+            custom_objects={
+                "policy": CustomMlpPolicy
+            }
+        )
+
+        # Make sure we loaded custom MLP policy
+        assert model.policy == CustomMlpPolicy
+        del model
 
     finally:
         if os.path.exists(model_fname):