+++ docs/misc/changelog.rst
@@ -30,6 +30,9 @@ Bug Fixes:
 
 Deprecations:
 ^^^^^^^^^^^^^
+- Models saved with cloudpickle format (stable-baselines<=2.7.0) are now 
+  deprecated in favor of zip-archive format for better support across
+  Python/Tensorflow versions. (@Miffyli)
 
 Others:
 ^^^^^^^
@@ -38,6 +41,8 @@ Others:
   to `stable_baselines.common.noise`. The API remains backward-compatible;
   for example `from stable_baselines.ddpg.noise import NormalActionNoise` is still
   okay. (@shwang)
+- **important change** Switch to using zip-archived JSON and Numpy `savez` for 
+  storing models for better support across library/Python verions. (@Miffyli)
 
 Documentation:
 ^^^^^^^^^^^^^^